Phase 8-9: Move serving patterns
- Move deployments/data_parallel → serving_patterns/data_parallel
- Move deployments/prefill_decode_disagg → serving_patterns/prefill_decode
- Rename builder_pd.py → builder.py
- Update all imports including release_tests
