+++ python/ray/llm/_internal/serve/serving_patterns/data_parallel/dp_server.py
@@ -5,7 +5,7 @@ from typing import Optional
 from ray import serve
 from ray.experimental.collective.util import get_address_and_port
 from ray.llm._internal.serve.core.configs.llm_config import LLMConfig
-from ray.llm._internal.serve.deployments.data_parallel.dp_rank_assigner import (
+from ray.llm._internal.serve.serving_patterns.data_parallel.dp_rank_assigner import (
     DPRankAssigner,
 )
 from ray.llm._internal.serve.core.server.builder import (
