+++ python/ray/llm/tests/serve/cpu/configs/test_multi_node_placement_groups.py
@@ -2,7 +2,7 @@ from typing import Any, Dict
 
 import pytest
 
-from ray.llm._internal.serve.deployments.data_parallel.dp_server import DPServer
+from ray.llm._internal.serve.serving_patterns.data_parallel.dp_server import DPServer
 from ray.llm._internal.serve.core.server.llm_server import LLMServer
 from ray.serve.llm import LLMConfig, ModelLoadingConfig
