+++ python/ray/llm/tests/serve/cpu/deployments/data_parallel/test_dp_server.py
@@ -4,7 +4,7 @@ from copy import deepcopy
 import pytest
 
 from ray.llm._internal.serve.core.configs.llm_config import LLMConfig
-from ray.llm._internal.serve.deployments.data_parallel.dp_server import DPServer
+from ray.llm._internal.serve.serving_patterns.data_parallel.dp_server import DPServer
 
 
 class TestGetDeploymentOptions:
