+++ python/ray/serve/llm/__init__.py
@@ -362,7 +362,7 @@ def build_pd_openai_app(pd_serving_args: dict) -> "Application":
     Returns:
         The configured Ray Serve Application router.
     """
-    from ray.llm._internal.serve.deployments.prefill_decode_disagg.builder_pd import (
+    from ray.llm._internal.serve.serving_patterns.prefill_decode.builder_pd import (
         build_pd_openai_app,
     )
