+++ release/llm_tests/serve/test_llm_serve_multi_node_integration.py
@@ -8,7 +8,7 @@ from ray.serve.llm import (
     ModelLoadingConfig,
     build_openai_app,
 )
-from ray.llm._internal.serve.deployments.data_parallel.dp_server import (
+from ray.llm._internal.serve.serving_patterns.data_parallel.dp_server import (
     build_dp_deployment,
 )
